AustralianElectricityMarkets.jl

A Julia package for interfacing with the Australian Energy Market Operator (AEMO) data archive.

Description

This package provides a set of tools to fetch, cache, and process data from AEMO's public data sources, particularly through the NEMWEB data archive. It allows users to read various datasets, and integrates with PowerSystems.jl for power system modeling and simulation.

Installation

To install the package, open the Julia REPL and run:

using Pkg
Pkg.add("AustralianElectricityMarkets")

Usage

Here is a basic example of how to use the package to read demand data:

using AustralianElectricityMarkets
using TidierDB

# Establish a database connection
db = connect(duckdb());

# Download the data from the monthly archive, saving them locally
# in parquet files
# Only the data requirements for a RegionalNetworkconfiguration are downloaded.
date_range = Date(2025, 1, 1):Date(2025, 1, 2)
fetch_table_data(date_range, RegionalNetworkConfiguration())

demand_df = read_demand(db)
println(demand_df)

And parsing the data into PowerSystems.jl

# Instantiate a System
sys = nem_system(db, RegionalNetworkConfiguration())

println(demand_df)

Features

Data Fetching: Fetch data from AEMO's NEMWEB data source for specified time ranges.

Data Processing: Read and process various NEMDB tables, including:

Interconnector data

Regional demand

Generation unit details

PowerSystems.jl Integration: Parse AEMO data into PowerSystems.jl models for simulation.

Data Sources

This package primarily interfaces with data from the Australian Energy Market Operator (AEMO). It relies on the NEMWEB data archives and data models.

License

This project is licensed under the MIT License.

Index

AustralianElectricityMarkets.HiveConfiguration

AustralianElectricityMarkets.RegionModel.RegionalNetworkConfiguration

AustralianElectricityMarkets.RegionModel._add_branches!

AustralianElectricityMarkets.RegionModel._add_buses!

AustralianElectricityMarkets.RegionModel._add_generation!

AustralianElectricityMarkets.RegionModel._add_loads!

AustralianElectricityMarkets.RegionModel.get_branch_dataframe

AustralianElectricityMarkets.RegionModel.get_bus_dataframe

AustralianElectricityMarkets.RegionModel.get_generators_dataframe

AustralianElectricityMarkets.RegionModel.get_load_dataframe

AustralianElectricityMarkets.RegionModel.nem_system

AustralianElectricityMarkets._add_demand_ts_to_components!

AustralianElectricityMarkets._add_renewable_ts_to_components!

AustralianElectricityMarkets._as_timearray

AustralianElectricityMarkets._filter_latest

AustralianElectricityMarkets._filter_latest

AustralianElectricityMarkets._parse_hive_root

AustralianElectricityMarkets.read_demand

AustralianElectricityMarkets.read_hive

AustralianElectricityMarkets.read_interconnectors

AustralianElectricityMarkets.read_units

AustralianElectricityMarkets.set_demand!

AustralianElectricityMarkets.set_market_bids!

AustralianElectricityMarkets.set_renewable_pv!

AustralianElectricityMarkets.set_renewable_wind!

AustralianElectricityMarkets.table_requirements

AustralianElectricityMarkets.HiveConfiguration — Type

HiveConfiguration

Configuration for accessing data.

Fields

hive_location::String: The directory where data is cached. Defaults to ~/.nemweb_cache.

filesystem::String: The filesystem to use for the cache. Defaults to "local". Can be s3, gs.

AustralianElectricityMarkets._add_demand_ts_to_components! — Method

_add_demand_ts_to_components!(sys, ts, type)

Adds demand time series data to the system components.

Arguments

sys: The PowerSystems.System object.

ts: A TimeArray of demand data.

type: The type of component to add the time series to.

AustralianElectricityMarkets._add_renewable_ts_to_components! — Method

_add_renewable_ts_to_components!(sys, ts, prime_mover)

Adds renewable generation time series data to the system components.

Arguments

sys: The PowerSystems.System object.

ts: A TimeArray of renewable generation data.

prime_mover: The prime mover type of the renewable generator.

AustralianElectricityMarkets._as_timearray — Method

_as_timearray(df, index, col, value)

Converts a DataFrame to a TimeArray.

Arguments

df: The input DataFrame.

index: The column to use as the timestamp.

col: The column to use for the column names of the TimeArray.

value: The column to use for the values of the TimeArray.

AustralianElectricityMarkets._filter_latest — Method

_filter_latest(table, key)

Helper function to filter for the most recent records in a table based on a given key.

Arguments

table: The TidierDB table to filter.

key: The column (as a Symbol) to use for determining the latest records.

AustralianElectricityMarkets._filter_latest — Method

_filter_latest(table, key=:archive_month)

Helper function to filter for the most recent records in a table.

Arguments

table: The table to filter.

key: The column to use for determining the latest records. Defaults to :archive_month.

AustralianElectricityMarkets._parse_hive_root — Method

_parse_hive_root(config::PyHiveConfiguration)

Construct the correct path to the Hive dataset based on the specified filesystem.

Arguments

config::PyHiveConfiguration: The configuration object containing filesystem and location details.

AustralianElectricityMarkets.read_demand — Method

read_demand(db; resolution::Dates.Period=Dates.Minute(5))

Read and process regional demand data from the database.

Arguments

db: The database connection.

resolution::Dates.Period: The time resolution to which the data should be floored. Defaults to 5 minutes.

Returns

A DataFrame with demand and renewable availability data, aggregated by the specified resolution.

Example

db = connect(duckdb())
demand_df = read_demand(db; resolution=Dates.Hour(1))
println(demand_df)

Row │ SETTLEMENTDATE REGIONID TOTALDEMAND DISPATCHABLEGENERATION DISPATCHABLELOAD NETINTERCHANGE │ Dates.DateTime String7 Float64 Float64 Float64 Float64 ───────┼────────────────────────────────────────────────────────────────────────────────────────────────────── 1 │ 2024-01-01T00:05:00 NSW1 6574.92 6721.88 0.0 146.96 2 │ 2024-01-01T00:05:00 QLD1 6228.31 5713.21 0.0 -515.1 3 │ 2024-01-01T00:05:00 SA1 1293.98 1116.68 0.0 -177.3 4 │ 2024-01-01T00:05:00 TAS1 1033.29 580.29 0.0 -453.0 5 │ 2024-01-01T00:05:00 VIC1 3977.1 5071.17 0.0 1094.07

AustralianElectricityMarkets.read_hive — Method

read_hive(db::TidierDB.DBInterface.Connection,table_name::Symbol; config::HiveConfiguration=CONFIG[])

Read a hive-partitioned parquet dataset into a TidierDB table.

Arguments

db::TidierDB.DBInterface.Connection: The database connection to use.

table_name::Symbol: The name of the table to read.

config::HiveConfiguration: The configuration to use. Defaults to CONFIG[].

AustralianElectricityMarkets.read_interconnectors — Method

read_interconnectors(db)

Reads and processes interconnector data from the database.

Arguments

db: The database connection.

Returns

A DataFrame containing the latest interconnector constraint data.

Example

db = connect(duckdb())
interconnectors_df = read_interconnectors(db)
println(interconnectors_df)

AustralianElectricityMarkets.read_units — Method

read_units(db)

Gathers and processes unit data from the database.

Arguments

db: The database connection.

Returns

A DataFrame containing detailed information about each generation unit.

Example

db = connect(duckdb())
units_df = read_units(db)
println(units_df)

AustralianElectricityMarkets.set_demand! — Method

set_demand!(sys, db, date_range; kwargs...)

Adds load time series data to the system from the database.

This function reads demand data for a specified date range, processes it into a time series, and attaches it to the PowerLoad components in the system.

Arguments

sys: The PowerSystems.System object.

db: The database connection.

date_range: A range of dates for which to fetch demand data.

kwargs: Additional keyword arguments passed to read_demand.

AustralianElectricityMarkets.set_market_bids! — Method

set_renewable_pv!(sys, db, date_range; kwargs...)

Adds Market bids time series data to the system.

This function reads solar availability data for a specified date range from the database, processes it into a time series, and attaches it to the RenewableDispatch components representing PV generators.

Arguments

sys: The PowerSystems.System object.

db: The database connection.

date_range: A range of dates for which to fetch the data.

kwargs: Additional keyword arguments passed to read_demand.

AustralianElectricityMarkets.set_renewable_pv! — Method

set_renewable_pv!(sys, db, date_range; kwargs...)

Adds photovoltaic (PV) renewable generation time series data to the system.

This function reads solar availability data for a specified date range from the database, processes it into a time series, and attaches it to the RenewableDispatch components representing PV generators.

Arguments

sys: The PowerSystems.System object.

db: The database connection.

date_range: A range of dates for which to fetch the data.

kwargs: Additional keyword arguments passed to read_demand.

AustralianElectricityMarkets.set_renewable_wind! — Method

set_renewable_wind!(sys, db, date_range; kwargs...)

Adds wind turbine renewable generation time series data to the system.

This function reads wind availability data for a specified date range from the database, processes it into a time series, and attaches it to the RenewableDispatch components representing wind turbines.

Arguments

sys: The PowerSystems.System object.

db: The database connection.

date_range: A range of dates for which to fetch the data.

kwargs: Additional keyword arguments passed to read_demand.

AustralianElectricityMarkets.RegionModel.RegionalNetworkConfiguration — Type

Each australian State (i.e. AEMO Region) is a combination of a load node and a generator node. The load nodes are connected.
between states by interconnectors

AustralianElectricityMarkets.RegionModel._add_branches! — Method

_add_branches!(sys, branch_df)

Adds branches to the system.

Arguments

sys: The PowerSystems.System object.

branch_df: A DataFrame of branch data.

AustralianElectricityMarkets.RegionModel._add_buses! — Method

_add_buses!(sys, bus_df)

Adds buses to the system.

Arguments

sys: The PowerSystems.System object.

bus_df: A DataFrame of bus data.

AustralianElectricityMarkets.RegionModel._add_generation! — Method

_add_generation!(sys, gen_df)

Adds generation units to the system.

Arguments

sys: The PowerSystems.System object.

gen_df: A DataFrame of generator data.

AustralianElectricityMarkets.RegionModel._add_loads! — Method

_add_loads!(sys, loads_df)

Adds loads to the system.

Arguments

sys: The PowerSystems.System object.

loads_df: A DataFrame of load data.

AustralianElectricityMarkets.RegionModel.get_branch_dataframe — Method

get_branch_dataframe(db)

Generates a DataFrame of branch information.

Arguments

db: The database connection.

Returns

A DataFrame containing branch details.

Example

db = connect(duckdb())
branch_df = get_branch_dataframe(db)
println(branch_df)

AustralianElectricityMarkets.RegionModel.get_bus_dataframe — Method

get_bus_dataframe(db)

Generates a DataFrame of bus information.

Arguments

db: The database connection.

Returns

A DataFrame containing bus details.

Example

db = connect(duckdb())
bus_df = get_bus_dataframe(db)
println(bus_df)

AustralianElectricityMarkets.RegionModel.get_generators_dataframe — Method

get_generators_dataframe(db)

Generates a DataFrame of generator information.

Arguments

db: The database connection.

Returns

A DataFrame containing generator details.

Example

db = connect(duckdb())
gen_df = get_generators_dataframe(db)
println(gen_df)

AustralianElectricityMarkets.RegionModel.get_load_dataframe — Method

get_load_dataframe(db)

Generates a DataFrame of load information.

Arguments

db: The database connection.

Returns

A DataFrame containing load details.

Example

db = connect(duckdb())
load_df = get_load_dataframe(db)
println(load_df)

AustralianElectricityMarkets.RegionModel.nem_system — Method

nem_system(db)

Assembles a PowerSystems.System object from the database.

Arguments

db: The database connection.

Returns

A PowerSystems.System object.

Example

db = connect(duckdb())
sys = nem_system(db)
println(sys)

AustralianElectricityMarkets.table_requirements — Method

table_requirements(::RegionalNetworkConfiguration)

:INTERCONNECTOR,
:INTERCONNECTORCONSTRAINT,
:DISPATCHREGIONSUM,
:DUDETAIL,
:DUDETAILSUMMARY,
:STATION,
:STATIONOPERATINGSTATUS,
:GENUNITS,
:DUALLOC,